src/third_party/blink/web_tests/fast/html/imports/import-style-tree-order-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M73, around March 2019. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 for more details.
PASS window.getComputedStyle(styleShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(linkShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(importShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(subimportShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(sharedImportShouldBeIgnored).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(styleShouldBeHiddenByDynamicallyAddedFollower).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(styleShouldNotBeHiddenByDynamicallyAddedLeader).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
